+++ .gitignore
@@ -4,4 +4,3 @@ root/
 root1/
 root2/
 Primate/
-gtree
