+++ cmd/gtree/main.go
@@ -2,14 +2,11 @@ package main
 
 import (
 	"fmt"
-	"io"
 	"log"
 	"os"
 	"strings"
 	"time"
 
-	"github.com/ddddddO/gtree"
-	"github.com/pkg/errors"
 	"github.com/urfave/cli/v2"
 )
 
@@ -187,94 +184,6 @@ func actionOutput(c *cli.Context) error {
 	return nil
 }
 
-type indentation uint
-
-const (
-	indentationUndefined indentation = iota
-	indentationTab
-	indentationTS
-	indentationFS
-)
-
-func decideIndentation(c *cli.Context) (indentation, error) {
-	if c.Bool("two-spaces") && c.Bool("four-spaces") {
-		return indentationUndefined, errors.New(`choose either "two-spaces(ts)" or "four-spaces(fs)".`)
-	}
-
-	indentation := indentationTab
-	if c.Bool("two-spaces") {
-		indentation = indentationTS
-	}
-	if c.Bool("four-spaces") {
-		indentation = indentationFS
-	}
-	return indentation, nil
-}
-
-type outputFormat uint
-
-const (
-	outputFormatUndefined outputFormat = iota
-	outputFormatStdout
-	outputFormatJSON
-	outputFormatYAML
-	outputFormatTOML
-)
-
-func decideOutputFormat(c *cli.Context) (outputFormat, error) {
-	if c.Bool("json") && c.Bool("yaml") && c.Bool("toml") {
-		return outputFormatUndefined, errors.New(`choose either "json(j)" or "yaml(y)" or "toml(t)".`)
-	}
-	if c.Bool("json") && c.Bool("yaml") {
-		return outputFormatUndefined, errors.New(`choose either "json(j)" or "yaml(y)".`)
-	}
-	if c.Bool("json") && c.Bool("toml") {
-		return outputFormatUndefined, errors.New(`choose either "json(j)" or "toml(t)".`)
-	}
-	if c.Bool("toml") && c.Bool("yaml") {
-		return outputFormatUndefined, errors.New(`choose either "toml(t)" or "yaml(y)".`)
-	}
-
-	outputFormat := outputFormatStdout
-	switch {
-	case c.Bool("json"):
-		outputFormat = outputFormatJSON
-	case c.Bool("yaml"):
-		outputFormat = outputFormatYAML
-	case c.Bool("toml"):
-		outputFormat = outputFormatTOML
-	}
-	return outputFormat, nil
-}
-
-func output(out io.Writer, in io.Reader, indentation indentation, outputFormat outputFormat, dryrun bool, extensions []string) error {
-	var options []gtree.OptFn
-
-	switch indentation {
-	case indentationTS:
-		options = append(options, gtree.WithIndentTwoSpaces())
-	case indentationFS:
-		options = append(options, gtree.WithIndentFourSpaces())
-	}
-
-	switch outputFormat {
-	case outputFormatJSON:
-		options = append(options, gtree.WithEncodeJSON())
-	case outputFormatYAML:
-		options = append(options, gtree.WithEncodeYAML())
-	case outputFormatTOML:
-		options = append(options, gtree.WithEncodeTOML())
-	}
-
-	if dryrun {
-		options = append(options, gtree.WithDryRun())
-	}
-
-	options = append(options, gtree.WithFileExtension(extensions))
-
-	return gtree.Output(out, in, options...)
-}
-
 func actionMkdir(c *cli.Context) error {
 	indentation, err := decideIndentation(c)
 	if err != nil {
@@ -305,21 +214,6 @@ func actionMkdir(c *cli.Context) error {
 	}
 
 	return nil
-}
-
-func mkdir(in io.Reader, indentation indentation, extensions []string) error {
-	var options []gtree.OptFn
-
-	switch indentation {
-	case indentationTS:
-		options = append(options, gtree.WithIndentTwoSpaces())
-	case indentationFS:
-		options = append(options, gtree.WithIndentFourSpaces())
-	}
-
-	options = append(options, gtree.WithFileExtension(extensions))
-
-	return gtree.Mkdir(in, options...)
 }
 
 var template = strings.TrimLeft(`
